Map is moved to home page.
