+++ client/src/index.jsx
@@ -31,13 +31,6 @@ class App extends React.Component {
       logged: false,
       image: null,
       photo: null,
-      mapCenter: {
-        lat: 34.7836966,
-        lng: -115.4089664
-      },
-      markers:[],
-      trailPopup: false
-
     }
 
     /*Bindings are set here.
@@ -45,72 +38,11 @@ class App extends React.Component {
     */
     this.submitImage = submitImage.bind(this);
     this.updateImageDisplay = updateImage.bind(this);
-    this.handleSearchBoxMounted = handleSearchBoxMounted.bind(this);
-    this.onMarkerClick = onMarkerClick.bind(this);
-    this.onMarkerClose = onMarkerClose.bind(this);
-    this.onDragEnd = onDragEnd.bind(this);
-    this.handleMapMounted = handleMapMounted.bind(this);
-    this.onPlacesChanged = handlePlacesChanged.bind(this);
-    this.trailClick = trailClick.bind(this);
   }
 
 //When the app mounts we are going to do the following actions. Get GPS location, find local trails, and load the current user.
 
   componentDidMount() {
-    this.input = document.querySelector('.input');
-    this.preview = document.querySelector('.preview');
-    //Note, right now the geolocation is just HTML5.
-    gps.getLocation()
-    .then(value => {
-      //First, we are going to get the location, then set it inside of an object.
-      let newObj = {
-        lat : value.coords.latitude,
-        lng : value.coords.longitude
-      }
-      //We then set that object as the mapCenter, which dictates where the Google Map locates at.
-      this.setState({mapCenter: newObj})
-    })
-    //Once we get the location of the user, we can then find trails in their area. We search using the MapCenter coordinates.
-    .then(() => {
-      return axios.get('/api/trails', {
-        params: {
-          lat: this.state.mapCenter.lat,
-          lng: this.state.mapCenter.lng,
-          radius: 10
-        }
-      });
-    })
-    .then(res => {
-      //Once we get trails from our particular area, we can then create markers on the map.
-      res.data.places.forEach((trail) => {
-        /*
-        We are going to set an array containing all of the markers.
-        The spread operator is used as a type of concat so that markers don't overwrite themselves.
-        */
-        const nextMarkers = [
-          ...this.state.markers,
-          {
-            /*
-            We set an object with some information that we are going to need later one.
-            In order to populate the map/list we use information passed in here.
-            */
-            position: {lat: trail.lat, lng: trail.lon},
-            name: trail.name,
-            city: trail.city,
-            state: trail.state,
-            //showInfo toggles the marker infowinow box. It is set to false right now, when the marker is clicked it will toggle to true.
-            showInfo: false,
-          },
-        ];
-        this.setState({
-          //The empty markers array in state is set to the populated nextMarkers array.
-          markers: nextMarkers,
-        });
-      });
-    })
-    .catch(err => {
-      console.log('oops, error in the trails call: ', err);
-    });
     axios.get('/api/currentUser')
       .then(res => {
         var email = res.data.email;
@@ -144,26 +76,6 @@ class App extends React.Component {
             <Upload submit={this.submitImage} update={this.updateImageDisplay}/>
           </Route>
         </Switch>
-        <div className = 'Gmap'  style={{
-            width: '700px',
-            height: '600px'
-          }}>
-          <Map
-            containerElement={< div style = {{width:100+'%', height:100+'%'}}/>}
-            mapElement={< div style = {{width:100+'%', height:100+'%'}}/>}
-            mapCenter={this.state.mapCenter}
-            onSearchBoxMounted={this.handleSearchBoxMounted}
-            markers = {this.state.markers}
-            onDragEnd={this.onDragEnd}
-            handleMapMounted={this.handleMapMounted}
-            onMarkerClose={this.onMarkerClose}
-            submit={this.submitImage}
-            update={this.updateImageDisplay}
-            onMarkerClick={this.onMarkerClick}
-            onPlacesChanged={this.onPlacesChanged}
-          />
-        </div>
-        <TrailList onClick={this.trailClick} markers={this.state.markers} />
       </div>
     )
   }
